関数を整理
